+++ ChangeLog.md
@@ -4,6 +4,7 @@
 * Put stackage snapshots JSON on S3 [#380](https://github.com/commercialhaskell/stack/issues/380)
 * Specifying flags for multiple packages [#335](https://github.com/commercialhaskell/stack/issues/335)
 * single test suite failure should show entire log [#388](https://github.com/commercialhaskell/stack/issues/388)
+* valid-wanted is a confusing option name [#386](https://github.com/commercialhaskell/stack/issues/386)
 
 ## 0.1.0.0